changed all pixel buffers to fixed 32 bit size. Fixes bars

+++ CrossProcess/BCImage.h
@@ -16,9 +16,9 @@
 {
     @private
     CGContextRef        _contextRef;        // CGContext for the image
-    NSUInteger*         _rawBytes;          // Raw pixel buffer
-    NSUInteger          _bufferSize;        // Pixel buffer size in bytes
-    NSUInteger          _rowBytes;          // Number of bytes per row (16-byte aligned)
+    uint32_t*           _rawBytes;          // Raw pixel buffer
+    uint32_t            _bufferSize;        // Pixel buffer size in bytes
+    uint32_t            _rowBytes;          // Number of bytes per row (16-byte aligned)
     
     CGSize              _size;              // Geometric size
 	UIImageOrientation	_orientation;		// Orientation
@@ -30,8 +30,8 @@
 }
 
 @property(nonatomic, readonly) CGContextRef			context;
-@property(nonatomic, readonly) NSUInteger*          bytes;
-@property(nonatomic, readonly) NSUInteger			bufferSize;
+@property(nonatomic, readonly) uint32_t*            bytes;
+@property(nonatomic, readonly) uint32_t             bufferSize;
 @property(nonatomic, readonly) CGSize				size;
 @property(nonatomic, readonly) CGImageRef			CGImageRef;
 @property(nonatomic, readonly) UIImageOrientation   orientation;
@@ -57,10 +57,10 @@
 #define GREEN_COMPONENT_RGBA(pixel)      (unsigned char)(*pixel >> 8)
 #define RED_COMPONENT_RGBA(pixel)        (unsigned char)(*pixel >> 0)
 
-#define SET_ALPHA_COMPONENT_RGBA(pixel, value)      *pixel = (*pixel & 0x00FFFFFF) | ((unsigned long)value << 24)
-#define SET_BLUE_COMPONENT_RGBA(pixel, value)       *pixel = (*pixel & 0xFF00FFFF) | ((unsigned long)value << 16)
-#define SET_GREEN_COMPONENT_RGBA(pixel, value)      *pixel = (*pixel & 0xFFFF00FF) | ((unsigned long)value << 8)
-#define SET_RED_COMPONENT_RGBA(pixel, value)        *pixel = (*pixel & 0xFFFFFF00) | ((unsigned long)value << 0)
+#define SET_ALPHA_COMPONENT_RGBA(pixel, value)      *pixel = (*pixel & 0x00FFFFFF) | (value << 24)
+#define SET_BLUE_COMPONENT_RGBA(pixel, value)       *pixel = (*pixel & 0xFF00FFFF) | (value << 16)
+#define SET_GREEN_COMPONENT_RGBA(pixel, value)      *pixel = (*pixel & 0xFFFF00FF) | (value << 8)
+#define SET_RED_COMPONENT_RGBA(pixel, value)        *pixel = (*pixel & 0xFFFFFF00) | (value << 0)
 
 // ARGB use this if CGImageAlphaInfo == kCGImageAlphaPremultipliedFirst
  
@@ -69,8 +69,8 @@
 #define RED_COMPONENT_ARGB(pixel)		(unsigned char)(*pixel >> 8)
 #define ALPHA_COMPONENT_ARGB(pixel)      (unsigned char)(*pixel >> 0)
 
-#define SET_BLUE_COMPONENT_ARGB(pixel, value)       *pixel = (*pixel & 0x00FFFFFF) | ((unsigned long)value << 24)
-#define SET_GREEN_COMPONENT_ARGB(pixel, value)      *pixel = (*pixel & 0xFF00FFFF) | ((unsigned long)value << 16)
-#define SET_RED_COMPONENT_ARGB(pixel, value)        *pixel = (*pixel & 0xFFFF00FF) | ((unsigned long)value << 8)
-#define SET_ALPHA_COMPONENT_ARGB(pixel, value)      *pixel = (*pixel & 0xFFFFFF00) | ((unsigned long)value << 0)
+#define SET_BLUE_COMPONENT_ARGB(pixel, value)       *pixel = (*pixel & 0x00FFFFFF) | (value << 24)
+#define SET_GREEN_COMPONENT_ARGB(pixel, value)      *pixel = (*pixel & 0xFF00FFFF) | (value << 16)
+#define SET_RED_COMPONENT_ARGB(pixel, value)        *pixel = (*pixel & 0xFFFF00FF) | (value << 8)
+#define SET_ALPHA_COMPONENT_ARGB(pixel, value)      *pixel = (*pixel & 0xFFFFFF00) | (value << 0)
